Basic RouterTranslator implementation
Implements router CRUD without gateway, routing, or SNAT functionality.

+++ midolman/src/test/scala/org/midonet/midolman/topology/devices/PortConversionTest.scala
@@ -137,7 +137,7 @@ class PortConversionTest extends FeatureSpec with Matchers {
             .setId(UUID.randomUUID.asProto)
             .setInboundFilterId(UUID.randomUUID.asProto)
             .setOutboundFilterId(UUID.randomUUID.asProto)
-            .setTunnelKey(random.nextLong())
+            .setTunnelKey(random.nextInt())
             .addPortGroupIds(UUID.randomUUID.asProto)
             .addPortGroupIds(UUID.randomUUID.asProto)
             .setPeerId(UUID.randomUUID.asProto)
